+++ lib/src/main/java/com/orange/ods/compose/component/progressindicator/OdsLinearProgressIndicator.kt
@@ -13,16 +13,21 @@ package com.orange.ods.compose.component.progressindicator
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
-import androidx.compose.material.Icon
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.width
+import androidx.compose.material.ButtonDefaults
 import androidx.compose.material.LinearProgressIndicator
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.ImageBitmap
 import androidx.compose.ui.graphics.painter.Painter
+import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.res.dimensionResource
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
@@ -32,6 +37,7 @@ import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.PreviewParameter
 import com.orange.ods.R
 import com.orange.ods.compose.component.OdsComposable
+import com.orange.ods.compose.component.content.OdsComponentIcon
 import com.orange.ods.compose.component.utilities.BasicPreviewParameterProvider
 import com.orange.ods.compose.component.utilities.Preview
 import com.orange.ods.compose.component.utilities.UiModePreviews
@@ -45,24 +51,22 @@ import com.orange.ods.extension.orElse
  * @see androidx.compose.material.LinearProgressIndicator
  *
  * @param modifier The modifier applied to this progress indicator
- * @param showCurrentValue To indicated if we have or not the current value
  * @param progress The value of this progress indicator, where 0.0 represents no progress and 1.0
  * represents full progress. Values outside of this range are coerced into the range. If set to `null`,
  * the progress indicator is indeterminate.
  * @param label The label displayed above the linear progress
  * @param icon The icon displayed above the linear progress
- * @param iconContentDescription The content description for the icon displayed above the linear progress
+ * @param showCurrentValue Indicates whether the current value is displayed
  */
 @OptIn(ExperimentalComposeUiApi::class)
 @Composable
 @OdsComposable
 fun OdsLinearProgressIndicator(
     modifier: Modifier = Modifier,
-    showCurrentValue: Boolean = false,
     progress: Float? = null,
     label: String? = null,
-    icon: Painter? = null,
-    iconContentDescription: String? = null
+    icon: OdsLinearProgressIndicatorIcon? = null,
+    showCurrentValue: Boolean = false
 ) {
     Column(
         modifier = modifier
@@ -74,12 +78,9 @@ fun OdsLinearProgressIndicator(
                 .padding(bottom = dimensionResource(id = R.dimen.spacing_xs)),
             horizontalArrangement = Arrangement.spacedBy(dimensionResource(id = R.dimen.spacing_s))
         ) {
-            icon?.let { painter ->
-                Icon(
-                    painter = painter, contentDescription = iconContentDescription,
-                    tint = OdsTheme.colors.onSurface
-                )
-            }
+            icon?.Content()
+            Spacer(Modifier.width(ButtonDefaults.IconSpacing))
+            
             if (label != null) {
                 Text(
                     text = label,
@@ -112,6 +113,42 @@ fun OdsLinearProgressIndicator(
     }
 }
 
+/**
+ * An icon in an [OdsLinearProgressIndicator].
+ * It is a non-clickable button.
+ */
+class OdsLinearProgressIndicatorIcon : OdsComponentIcon {
+
+    /**
+     * Creates an instance of [OdsLinearProgressIndicatorIcon].
+     *
+     * @param painter Painter of the icon.
+     * @param contentDescription The content description associated to this [OdsLinearProgressIndicatorIcon].
+     */
+    constructor(painter: Painter, contentDescription: String) : super(painter as Any, contentDescription)
+
+    /**
+     * Creates an instance of [OdsLinearProgressIndicatorIcon].
+     *
+     * @param imageVector Image vector of the icon.
+     * @param contentDescription The content description associated to this [OdsLinearProgressIndicatorIcon].
+     */
+    constructor(imageVector: ImageVector, contentDescription: String) : super(imageVector as Any, contentDescription)
+
+    /**
+     * Creates an instance of [OdsLinearProgressIndicatorIcon].
+     *
+     * @param bitmap Image bitmap of the icon.
+     * @param contentDescription The content description associated to this [OdsLinearProgressIndicatorIcon].
+     */
+    constructor(bitmap: ImageBitmap, contentDescription: String) : super(bitmap as Any, contentDescription)
+
+    @Composable
+    override fun Content(modifier: Modifier) {
+        super.Content(modifier = modifier.size(ButtonDefaults.IconSize))
+    }
+}
+
 @UiModePreviews.Default
 @Composable
 private fun PreviewOdsLinearProgressIndicator(@PreviewParameter(OdsLinearProgressIndicatorPreviewParameterProvider::class) parameter: OdsLinearProgressIndicatorPreviewParameter) =
@@ -119,7 +156,7 @@ private fun PreviewOdsLinearProgressIndicator(@PreviewParameter(OdsLinearProgres
         with(parameter) {
             OdsLinearProgressIndicator(
                 progress = progress,
-                icon = iconRes?.let { painterResource(id = it) },
+                icon = iconRes?.let { OdsLinearProgressIndicatorIcon(painterResource(id = it), "") },
                 label = label,
                 showCurrentValue = showCurrentValue
             )